moved trial etl-specific parser
